Lent/requirements.txt
numpy==1.24.3
einops==0.6.1
h5py==3.7.0
matplotlib-label-lines==0.6.0
seaborn==0.12.2
wandb==0.15.0
pyYAML==6.0
argparse==1.4.0
tqdm==4.65.0
matplotlib==3.7.1
pandas==2.0.1
scipy==1.10.1
hydra-core==1.3.2
torch-intermediate-layer-getter==0.1.post1
scikit-learn==1.2.2